requirements.txt
aiohttp==3.9.1
asyncio==3.4.3
PyGithub==2.1.1
python-telegram-bot==20.7
discord-webhook==1.3.0
flask==3.0.0
flask-socketio==5.3.6
sqlalchemy==2.0.23
pyyaml==6.0.1
rich==13.7.0
colorama==0.4.6
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
dnspython==2.4.2
shodan==1.31.0
censys==2.2.9
schedule==1.2.1
cryptography==41.0.7
gevent==23.9.1
greenlet==3.0.1
eventlet==0.33.3